Fix the parser for use with v18 of autogtp (#40)
The EVENT parser is pretty fragile in the sense that if the initial text sent by autogtp changes at all, it can fail in a mysterious ways.
I updated the regexp to accept the new text and still be backward compatible with old versions, but it could fail again in the future
if the autogtp start text changes again.

I bumped the version to 1.2.0 and also updated to gradle 6.2 and a newer version of the fatjar plugin. You can now build with java 11.